Moved higher order components to src
